+++ CHANGELOG.md
@@ -5,7 +5,43 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Development]
+
+### [Development]
+
+### [0.1.0] - 2023-10-21
+
+### Added
+
+- Create Lectern View
+- Add multipliers in board
+- Add multipliers in board tests
+
+### Changed
+
+- Solve problems with change tiles in main
+- Modify tests in main
+
+### Issues
+
+- Problems with put word in main
+
+### [0.0.9] - 2023-10-12
+
+### Added
+
+- Create Main and his tests
+
+### Changed
+
+- Modify board class
+
+- Modify tests in board
+
+- Modify tests in scrabblegame
+
+### Issues
+
+- Problems with change tiles and put word in main and his tests
 
 ### [0.0.8] - 2023-10-05